pyproject.toml
[project]
name = "genai-3d-vr-ar"
version = "1.0.0"
description = "Production-ready GenAI 3D Virtual Reality and Augmented Reality platform with WatsonX.ai integration"
authors = [
    {name = "Ruslan Magana", email = "contact@ruslanmv.com"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
keywords = [
    "ai",
    "genai",
    "virtual-reality",
    "augmented-reality",
    "3d-generation",
    "watsonx",
    "stable-diffusion",
    "unreal-engine",
    "vr",
    "ar"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "gradio>=4.44.0",
    "python-dotenv>=1.0.0",
    "ibm-watson-machine-learning>=1.0.355",
    "diffusers>=0.30.0",
    "transformers>=4.44.0",
    "torch>=2.4.0",
    "accelerate>=0.34.0",
    "safetensors>=0.4.5",
    "Pillow>=10.4.0",
    "numpy>=1.26.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "black>=24.8.0",
    "mypy>=1.11.0",
    "isort>=5.13.0",
    "pre-commit>=3.8.0",
]

test = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "responses>=0.25.0",
]

docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
]

[project.urls]
Homepage = "https://ruslanmv.com"
Documentation = "https://github.com/ruslanmv/GenAI-3D-Virtual-Reality-Augmented#readme"
Repository = "https://github.com/ruslanmv/GenAI-3D-Virtual-Reality-Augmented"
Issues = "https://github.com/ruslanmv/GenAI-3D-Virtual-Reality-Augmented/issues"

[project.scripts]
genai-3d-vr = "genai_3d_vr_ar.app:main"
genai-env-gen = "genai_3d_vr_ar.generate_environment:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["genai_3d_vr_ar"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "black>=24.8.0",
    "mypy>=1.11.0",
]

[tool.uv.sources]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "ibm_watson_machine_learning.*",
    "diffusers.*",
    "transformers.*",
    "gradio.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --cov=genai_3d_vr_ar --cov-report=term-missing --cov-report=html"
testpaths = [
    "tests",
]
pythonpath = [
    "."
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["genai_3d_vr_ar"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if False:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
